+++ tomcat/src/main/java/org/apache/coyote/config/AppConfig.java
@@ -1,8 +1,8 @@
 package org.apache.coyote.config;
 
 import org.apache.coyote.render.PageRenderer;
-import org.apache.coyote.render.UserLoginHandler;
-import org.apache.coyote.render.UserRegisterHandler;
+import org.apache.coyote.handler.UserLoginHandler;
+import org.apache.coyote.handler.UserRegisterHandler;
 import org.apache.coyote.router.RequestRouter;
 
 public class AppConfig {
